Added UniverseSyncService that gets regions, systems and locations, added worker models and edited stuff to use that

+++ src/EveOnTrader.Core/Models/MarketLocation.cs
@@ -3,7 +3,7 @@
 public class MarketLocation
 {
     public long LocationId { get; set; }
-    public int SolarSystemId { get; set; }
+    public long SolarSystemId { get; set; }
     public string Name { get; set; } = "";
     public KindValue Kind { get; set; }
     public bool? HasPublicDocking { get; set; }
