+++ src/EveOnTrader.Core/Models/SolarSystem.cs
@@ -2,8 +2,8 @@
 
 public class SolarSystem
 {
-    public int SolarSystemId { get; set; }
-    public int RegionId { get; set; }
+    public long SolarSystemId { get; set; }
+    public long RegionId { get; set; }
     public string Name { get; set; } = "";
     public double SecurityStatus { get; set; }
 }
